refactor: clean up pre-aggregate files
